Refactoring code for new reading Removed code that runs after a new reading has been received from the express model, and added it as a sequelize hook instead.
